move sidecar logging functions to separate module
