+++ src/sidecar/sidecarManager.ts
@@ -537,67 +537,6 @@ export class SidecarManager {
   }
 
   /**
-   * Set up tailing the sidecar log file onto the vscode output channel.
-   **/
-  private startTailingSidecarLogs() {
-    // Create sidecar's log file if it doesn't exist so that we can
-    // start tailing it right away before the sidecar process may exist.
-
-    const sidecarLogfilePath = getSidecarLogfilePath();
-    try {
-      fs.accessSync(sidecarLogfilePath);
-    } catch {
-      fs.writeFileSync(sidecarLogfilePath, "");
-    }
-
-    try {
-      this.logTailer = new Tail(sidecarLogfilePath);
-    } catch (e) {
-      SIDECAR_OUTPUT_CHANNEL.appendLine(
-        `Failed to tail sidecar log file "${sidecarLogfilePath}": ${e}`,
-      );
-      return;
-    }
-
-    SIDECAR_OUTPUT_CHANNEL.appendLine(
-      `Tailing the extension's sidecar logs from "${sidecarLogfilePath}" ...`,
-    );
-
-    // Take note of the start of exception lines in the log file, show as toast (if user has allowed via config)
-    // Define a regex pattern to find "ERROR", a parenthesized thread name, and capture everything after it
-    const regex = /ERROR.*\(([^)]+)\)\s*(.*)$/;
-
-    this.logTailer.on("line", (data: any) => {
-      const line: string = data.toString();
-      const errorMatch = line.match(regex);
-      if (errorMatch) {
-        const config = vscode.workspace.getConfiguration();
-        const notifySidecarExceptions = config.get(
-          "confluent.debugging.showSidecarExceptions",
-          false,
-        );
-        if (notifySidecarExceptions) {
-          showErrorNotificationWithButtons(`[Debugging] Sidecar error: ${errorMatch[2]}`, {
-            "Open Sidecar Logs": () =>
-              vscode.commands.executeCommand("confluent.showSidecarOutputChannel"),
-            "Open Settings": () =>
-              vscode.commands.executeCommand(
-                "workbench.action.openSettings",
-                "@id:confluent.debugging.showSidecarExceptions",
-              ),
-          });
-        }
-      }
-
-      appendSidecarLogToOutputChannel(line);
-    });
-
-    this.logTailer.on("error", (data: any) => {
-      OUTPUT_CHANNEL.error(`Error tailing sidecar log: ${data.toString()}`);
-    });
-  }
-
-  /**
    * Get the auth token secret from the storage manager. Returns empty string if none found.
    **/
   async getAuthTokenFromSecretStore(): Promise<string> {
@@ -761,74 +700,6 @@ export function safeKill(process_id: number, signal: NodeJS.Signals | 0 = "SIGTE
     process.kill(process_id, signal);
   } catch (e) {
     logger.error(`Failed to deliver signal ${signal} to process ${process_id}: ${e}`);
-  }
-}
-
-/** @see https://quarkus.io/guides/logging#logging-format */
-interface SidecarLogFormat {
-  timestamp: string;
-  sequence: number;
-  loggerClassName: string; // usually "org.jboss.logging.Logger"
-  loggerName: string;
-  level: string;
-  message: string;
-  threadName: string;
-  threadId: number;
-  mdc: Record<string, unknown>;
-  ndc: string;
-  hostName: string;
-  processName: string;
-  processId: number;
-}
-
-/**
- * Parse and append a sidecar log line to the {@link SIDECAR_OUTPUT_CHANNEL output channel} based on
- * its `level`.
- */
-export function appendSidecarLogToOutputChannel(line: string) {
-  // DEBUGGING: uncomment to see raw log lines in the output channel
-  // SIDECAR_OUTPUT_CHANNEL.trace(line);
-
-  let log: SidecarLogFormat;
-  try {
-    log = JSON.parse(line) as SidecarLogFormat;
-  } catch (e) {
-    if (e instanceof Error) {
-      OUTPUT_CHANNEL.error(`Failed to parse sidecar log line: ${e.message}\n\t${line}`);
-    }
-    return;
-  }
-  if (!(log.level && log.loggerName && log.message)) {
-    // log the raw object at `info` level:
-    SIDECAR_OUTPUT_CHANNEL.appendLine(line);
-    return;
-  }
-
-  let logMsg = `[${log.loggerName}] ${log.message}`;
-
-  const logArgs = [];
-  if (log.mdc && Object.keys(log.mdc).length > 0) {
-    logArgs.push(log.mdc);
-  }
-
-  switch (log.level) {
-    case "DEBUG":
-      SIDECAR_OUTPUT_CHANNEL.debug(logMsg, ...logArgs);
-      break;
-    case "INFO":
-      SIDECAR_OUTPUT_CHANNEL.info(logMsg, ...logArgs);
-      break;
-    case "WARN":
-      SIDECAR_OUTPUT_CHANNEL.warn(logMsg, ...logArgs);
-      break;
-    case "ERROR":
-      SIDECAR_OUTPUT_CHANNEL.error(logMsg, ...logArgs);
-      break;
-    default:
-      // still shows up as `info` in the output channel
-      SIDECAR_OUTPUT_CHANNEL.appendLine(
-        `[${log.level}] ${logMsg} ${logArgs.length > 0 ? JSON.stringify(logArgs) : ""}`.trim(),
-      );
   }
 }
